renamed packages (core -> entities, correlogram -> core)
