+++ core/lambda_calc.py
@@ -1,5 +1,3 @@
-from core.image import Image
-
 def vertical_lambda(pixel, distance): # like horizontal_lambda but in an imperative style.
   sum = 0
   for i in range(0, 2 * distance - 1): # 0 ... 2k-2 times
