+++ main.py
@@ -1,6 +1,6 @@
 #! /Users/yishaizehavi/.virtualenvs/cv/bin/python3
 
-from correlogram import specific_correlogram
+from core.correlogram import autocorrelogram
 from utils.colors import get_unique_colors
 from adapters.cv_image import OpenCVImage
 from utils.log import log_time
@@ -11,7 +11,7 @@ def main():
   colors = get_unique_colors(luffy)
 
   for color in colors:
-    print(specific_correlogram(luffy, color, color, 8))
+    print(autocorrelogram(luffy, color, 8))
 
 if __name__ == "__main__":
   main()
