+++ utils/colors.py
@@ -1,5 +1,5 @@
 import numpy as np
-from core.image import Image
+from entities.image import Image
 
 def get_unique_colors(image: Image):
     flat_image = image.source.reshape(-1)
